requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
load-dotenv==0.1.0
MarkupSafe==3.0.2
packaging==24.2
paho-mqtt==2.1.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-etcd==0.4.5
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
